+++ src/main/java/net/hawkelele/quickinfopanel/QuickInfoPanel.java
@@ -1,14 +1,22 @@
 package net.hawkelele.quickinfopanel;
 
 import net.fabricmc.api.ClientModInitializer;
-import net.hawkelele.quickinfopanel.events.command.ExtraCommand;
-import net.hawkelele.quickinfopanel.events.command.PositionCommand;
-import net.hawkelele.quickinfopanel.events.keybinds.SecondaryToggleKeybindPressed;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.message.v1.ClientReceiveMessageEvents;
+import net.fabricmc.fabric.api.client.rendering.v1.hud.HudElementRegistry;
+import net.fabricmc.fabric.api.client.rendering.v1.hud.VanillaHudElements;
+import net.hawkelele.quickinfopanel.commands.ExtraCommand;
+import net.hawkelele.quickinfopanel.commands.PositionCommand;
+import net.hawkelele.quickinfopanel.events.ActionBarCallback;
+import net.hawkelele.quickinfopanel.input.SecondaryPanelToggleKeyPress;
+import net.hawkelele.quickinfopanel.gui.Panel;
+import net.hawkelele.quickinfopanel.providers.client.OverlayMessageStatusProvider;
 import net.hawkelele.quickinfopanel.registry.CommandRegistry;
 import net.hawkelele.quickinfopanel.registry.KeybindRegistry;
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.resources.Identifier;
-import net.hawkelele.quickinfopanel.events.keybinds.MainToggleKeybindPressed;
+import net.hawkelele.quickinfopanel.input.PanelToggleKeyPress;
 
 
 public class QuickInfoPanel implements ClientModInitializer {
@@ -17,11 +25,23 @@ public class QuickInfoPanel implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
+        Panel panel = new Panel();
+
+        // Attach our rendering code to before the chat hud layer. Our layer will render right before the chat. The API will take care of z spacing.
+        HudElementRegistry.attachElementBefore(VanillaHudElements.CHAT, Identifier.fromNamespaceAndPath(MOD_ID, "qip"), panel);
+
+
+        ClientReceiveMessageEvents.GAME.register((message, overlay) -> {
+                panel.hideFor(60);
+
+        });
+
+        ClientTickEvents.END_CLIENT_TICK.register((minecraft) -> panel.tick());
 
 
         KeybindRegistry.register(
-                new MainToggleKeybindPressed(),
-                new SecondaryToggleKeybindPressed()
+                new PanelToggleKeyPress(),
+                new SecondaryPanelToggleKeyPress()
         );
 
         CommandRegistry.register(