+++ src/main/java/net/hawkelele/quickinfopanel/gui/layouts/RootLayout.java
@@ -1,7 +1,8 @@
-package net.hawkelele.quickinfopanel.gui;
+package net.hawkelele.quickinfopanel.gui.layouts;
 
 import net.hawkelele.quickinfopanel.config.Config;
 import net.hawkelele.quickinfopanel.gui.core.elements.Layout;
+import net.hawkelele.quickinfopanel.providers.client.OverlayMessageStatusProvider;
 import net.hawkelele.quickinfopanel.registry.LayoutRegistry;
 import net.minecraft.client.Minecraft;
 
@@ -14,6 +15,7 @@ public class RootLayout extends Layout {
                 || client.player == null
                 || client.level == null
                 || client.debugEntries.isOverlayVisible()
+                || OverlayMessageStatusProvider.isDisplaying()
                 || !Config.getInstance().settings().displayMainPanel;
     }