+++ src/main/java/net/hawkelele/quickinfopanel/input/PanelToggleKeyPress.java
@@ -1,4 +1,4 @@
-package net.hawkelele.quickinfopanel.events.keybinds;
+package net.hawkelele.quickinfopanel.input;
 
 import com.mojang.blaze3d.platform.InputConstants;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
@@ -9,7 +9,7 @@ import net.hawkelele.quickinfopanel.events.EventHandler;
 import net.minecraft.client.KeyMapping;
 import org.lwjgl.glfw.GLFW;
 
-public class MainToggleKeybindPressed extends EventHandler<ClientTickEvents.EndTick> {
+public class PanelToggleKeyPress extends EventHandler<ClientTickEvents.EndTick> {
     private static final KeyMapping keyBinding = KeyMappingHelper.registerKeyMapping(new KeyMapping("key.quickinfopanel.toggle", // The translation key of the keybinding's name
             InputConstants.Type.KEYSYM, // The type of the keybinding, KEYSYM for keyboard, MOUSE for mouse.
             GLFW.GLFW_KEY_B, // The keycode of the key